+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `isPostalCode()`-validation. With the new `validator.js` version `8.1.0` validation of postal codes can be performed for several different locales.
 - Norwegian locales `nb-NO` and `nn-NO` are now supported for `isAlpha()` and `isAlphanumeric()` validations.
 - `isMobilePhone()` now supports locales for Kenya `en-KE`, Rwanda `en-RW`, Tanzania `en-TZ`, Uganda `en-UG`, Indonesia `id-ID`, Iran `fa-IR`, and Lithuania `lt-LT`.
+- Additionally to checking if a certain value/object is an element of an iterable, 
+`isIn()` can now validate if a given object has a field with a certain value. ([Issue 5](https://github.com/markusbohl/fluent-ts-validator/issues/5))
 
 ### Changed
 - `dependencies` have been updated as follows: